# Conflicts:
#	frontend/src/components/Reaction/Scrap.tsx
#	frontend/src/pages/StudylogPage/Content.tsx
#	frontend/src/pages/StudylogPage/index.js

+++ backend/src/main/java/wooteco/prolog/ability/application/StudylogAbilityService.java
@@ -2,6 +2,7 @@ package wooteco.prolog.ability.application;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 import org.springframework.context.event.EventListener;
 import org.springframework.data.domain.Page;
@@ -17,6 +18,7 @@ import wooteco.prolog.ability.domain.repository.StudylogAbilityRepository;
 import wooteco.prolog.common.PageableResponse;
 import wooteco.prolog.member.application.MemberService;
 import wooteco.prolog.member.domain.Member;
+import wooteco.prolog.report.application.dto.StudylogPeriodRequest;
 import wooteco.prolog.studylog.application.StudylogService;
 import wooteco.prolog.studylog.domain.Studylog;
 import wooteco.prolog.studylog.event.StudylogDeleteEvent;
@@ -40,37 +42,46 @@ public class StudylogAbilityService {
     }
 
     @Transactional
-    public List<AbilityResponse> updateStudylogAbilities(Long memberId, Long studylogId, StudylogAbilityRequest studylogAbilityRequest) {
+    public List<AbilityResponse> updateStudylogAbilities(Long memberId, Long studylogId,
+                                                         StudylogAbilityRequest studylogAbilityRequest) {
+        Set<Ability> abilities = abilityService.findByIdIn(memberId,
+            studylogAbilityRequest.getAbilities());
+
+        // 자식 역량이 있는데 부모 역량이 있는 경우 예외처리
+        if (hasChildAndParentAbility(abilities)) {
+            throw new IllegalArgumentException("자식 역량이 존재하는 경우 부모 역량을 선택할 수 없습니다.");
+        }
+
         Studylog studylog = studylogService.findStudylogById(studylogId);
         studylog.validateBelongTo(memberId);
-
-        List<Ability> abilities = abilityService.findByIdIn(memberId, studylogAbilityRequest.getAbilities());
-        // 자식 역량이 있는데 부모 역량이 있는 경우 예외처리
-        abilities.stream()
-            .filter(it -> !it.isParent())
-            .filter(it -> abilities.contains(it.getParent()))
-            .findFirst()
-            .ifPresent(it -> {
-                throw new IllegalArgumentException("자식 역량이 존재하는 경우 부모 역량을 선택할 수 없습니다.");
-            });
 
         List<StudylogAbility> studylogAbilities = abilities.stream()
             .map(it -> new StudylogAbility(memberId, it, studylog))
             .collect(Collectors.toList());
 
         studylogAbilityRepository.deleteByStudylogId(studylogId);
-        List<StudylogAbility> persistStudylogAbilities = studylogAbilityRepository.saveAll(studylogAbilities);
+        List<StudylogAbility> persistStudylogAbilities = studylogAbilityRepository.saveAll(
+            studylogAbilities);
 
         return persistStudylogAbilities.stream()
             .map(it -> AbilityResponse.of(it.getAbility()))
             .collect(Collectors.toList());
     }
 
-    public PageableResponse<AbilityStudylogResponse> findAbilityStudylogsByAbilityIds(String username, List<Long> abilityIds, Pageable pageable) {
+    private boolean hasChildAndParentAbility(Set<Ability> abilities) {
+        return abilities.stream()
+            .anyMatch(ability -> !ability.isParent() &&
+                abilities.contains(ability.getParent()));
+    }
+
+    public PageableResponse<AbilityStudylogResponse> findAbilityStudylogsByAbilityIds(
+        String username, List<Long> abilityIds, Pageable pageable) {
         if (abilityIds != null && !abilityIds.isEmpty()) {
 
-            Page<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByAbilityIdIn(abilityIds, pageable);
-            List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(studylogAbilities.getContent());
+            Page<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByAbilityIdIn(
+                abilityIds, pageable);
+            List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(
+                studylogAbilities.getContent());
             return PageableResponse.of(abilityStudylogResponses, studylogAbilities);
         }
 
@@ -79,34 +90,52 @@ public class StudylogAbilityService {
             .map(Studylog::getId)
             .collect(Collectors.toList());
 
-        List<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByStudylogIdIn(studylogIds);
+        List<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByStudylogIdIn(
+            studylogIds);
 
-        List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(studylogs.getContent(), studylogAbilities);
+        List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(
+            studylogs.getContent(), studylogAbilities);
         return PageableResponse.of(abilityStudylogResponses, studylogs);
     }
 
-    public PageableResponse<AbilityStudylogResponse> findAbilityStudylogsMappingOnlyByAbilityIds(String username, List<Long> abilityIds, Pageable pageable) {
+    public PageableResponse<AbilityStudylogResponse> findAbilityStudylogsMappingOnlyByAbilityIds(
+        String username, List<Long> abilityIds, Pageable pageable) {
         if (abilityIds != null && !abilityIds.isEmpty()) {
-            Page<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByAbilityIdIn(abilityIds, pageable);
-            List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(studylogAbilities.getContent());
+            Page<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByAbilityIdIn(
+                abilityIds, pageable);
+            List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(
+                studylogAbilities.getContent());
             return PageableResponse.of(abilityStudylogResponses, studylogAbilities);
         }
 
         Member member = memberService.findByUsername(username);
 
-        Page<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByMemberId(member.getId(), pageable);
+        Page<StudylogAbility> studylogAbilities = studylogAbilityRepository.findByMemberId(
+            member.getId(), pageable);
 
-        List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(studylogAbilities.getContent());
+        List<AbilityStudylogResponse> abilityStudylogResponses = AbilityStudylogResponse.listOf(
+            studylogAbilities.getContent());
         return PageableResponse.of(abilityStudylogResponses, studylogAbilities);
     }
 
-    public List<StudylogAbility> findStudylogAbilitiesInPeriod(Long memberId, LocalDate startDate, LocalDate endDate) {
-        List<Studylog> studylogs = studylogService.findStudylogsInPeriod(memberId, startDate, endDate);
-        return studylogAbilityRepository.findByStudylogIdIn(studylogs.stream().map(Studylog::getId).collect(Collectors.toList()));
+    public List<StudylogAbility> findStudylogAbilitiesInPeriod(Long memberId, LocalDate startDate,
+                                                               LocalDate endDate) {
+        List<Studylog> studylogs = studylogService.findStudylogsInPeriod(memberId, startDate,
+            endDate);
+        return studylogAbilityRepository.findByStudylogIdIn(
+            studylogs.stream().map(Studylog::getId).collect(Collectors.toList()));
     }
 
     @EventListener
     public void onStudylogDeleteEvent(StudylogDeleteEvent event) {
         studylogAbilityRepository.deleteByStudylogId(event.getStudylogId());
+    }
+
+    public List<AbilityStudylogResponse> getStudylogsByDate(Long memberId,
+                                                            StudylogPeriodRequest studylogPeriodRequest) {
+        List<StudylogAbility> studylogAbilities = findStudylogAbilitiesInPeriod(memberId, LocalDate.parse(studylogPeriodRequest.getStartDate()),
+                        LocalDate.parse(studylogPeriodRequest.getEndDate()));
+
+        return AbilityStudylogResponse.listOf(studylogAbilities);
     }
 }
